resolved circular dependencies sent example message to frontend send message on update
