+++ internal/config/pool.go
@@ -1,0 +1,8 @@
+package config
+
+import (
+	"github.com/CHainGate/backend/internal/model"
+	"github.com/google/uuid"
+)
+
+var Pools = make(map[uuid.UUID]*model.Pool)
